Differentiated scroll from touchmove event
When navigation is scrolled, the touchnove event may be triggered. Now this has been fixed using the Rectangle class and the scrollControl(Set) varaibles. When touchmoves in Y-axis, until touchend before nav can listen for touchmove on X-axis. So if touchmoves initially in Y-axis and branches to X-axis nav doesn't listen until another touchend and a initial touchmove of X-axis